Refactor AzurePublisher to enable usage in downstream publishers (#8507)
Few refactoring changes to AzurePublisher to enable usage in downstream publishers:

Refactor the logic in AzurePublisher to a new public AzurePublishingContext class.
refactor code in AzurePublisher.PublishAsync to AzurePublishingContext and capture bicep compilation data that is thrown away in the existing AzurePublisher.PublishAsync method

* Restructure AzurePublisher and add AzurePublishingContext

* Few more edits and add an playgroud app for AzurePublisher

* Move SaveToDisk to Publisher

* Edit Playground aspire starter app to use AspireProjectOrPackageReference and call AddAzurePublisher

* Remove AzurePublisher playground app

* Refactor AzurePublishingContext to be public.

* Fix up .sln file

---------
